customer-complaints-etl/requirements.txt
pandas==2.1.4
psycopg2-binary==2.9.9
faker==20.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
numpy==1.26.2
